[Discussion] Intentional to mask `error.cause`?
Thought it was easier to create a discussion with my proposed change here.

I was debugging some errors in production and noticed that my `cause`-tree was missing and I think I have drilled it down to this line of code.

Often times, libraries wrap errors in their own errors so they are homogenic (we do this in tRPC by wrapping all in `TRPCError` with the original error as the `cause` and Apollo Server does something similar with GraphQL Errors IIRC). 

For debugging however, the original error may contain useful properties, which seem to be hidden here.

I can work around this, but wanted to ask some more motivation on how come this is and if this is intentional? 

It might be me misinterpreting how to use things